refactor(record): reorganize match constants and update usage in components
